Modify CI pipeline

+++ hack/release-helm-chart.sh
@@ -22,19 +22,29 @@ helm::package() {
 
 chartmuseum::push() {
   local version="$(chart::version)"
-  if [ -n "${SNAPSHOT}" ]; then
-    version="${version}-SNAPSHOT"
 
-    curl ${CHARTMUSEUM_ENDPOINT}/api/charts/organizer-bot/${version} \
-      -u "${CHARTMUSEUM_AUTH_USER}:${CHARTMUSEUM_AUTH_PASS}" \
-      -XDELETE
+  local exists=$(curl ${CHARTMUSEUM_ENDPOINT}/api/charts/organizer-bot/${version} \
+      -u "${CHARTMUSEUM_AUTH_USER}:${CHARTMUSEUM_AUTH_PASS}" | jq -r '.name')
+
+  if [ "$exists" != "null" ]; then
+    if [ -n "${SNAPSHOT}" ]; then
+      version="${version}-SNAPSHOT"
+
+      curl ${CHARTMUSEUM_ENDPOINT}/api/charts/organizer-bot/${version} \
+        -u "${CHARTMUSEUM_AUTH_USER}:${CHARTMUSEUM_AUTH_PASS}" \
+        -XDELETE
+    else
+      echo "Chart in version ${version} already released"
+      return 0
+    fi
   fi
 
   local -r post_resp=$(curl ${CHARTMUSEUM_ENDPOINT}/api/charts \
     -u "${CHARTMUSEUM_AUTH_USER}:${CHARTMUSEUM_AUTH_PASS}" \
     --data-binary "@organizer-bot-${version}.tgz")
-  echo $post_resp
-  echo "$post_resp" | jq -er ".saved"
+  echo "$post_resp" | jq -er ".saved" > /dev/null
+
+  echo "Chart ${version} was successfully uploaded!"
 }
 
 main() {
